Added lots of needed features and bug fixes.
the game now has an android version fixed saving of the local scoreboard attempted to add a local scoreboard. does not seam to save properly added the ability to pause the game. for this, use p attempted to fix the hint bugs. needs further testing. The game no longer adds duplicate numbers to the list. Added the game clock. needs testing for bugs